parametro i de insereEntre modificado para int para comparacao com unsigned int

+++ Patricia.h
@@ -36,7 +36,7 @@ TipoPatNo* CriaNoExt (char *k, int idDoc);
 float RetornaPesoTermo (char *k, TipoPatNo *t, int idDoc);
 
 /* Funções para inserir uma palavra na árvore Patricia */
-TipoPatNo* InsereEntre (char *k, TipoPatNo **t, short i, int idDoc, char Caractere);
+TipoPatNo* InsereEntre (char *k, TipoPatNo **t, int i, int idDoc, char Caractere);
 TipoPatNo* InserePatricia (char *k, TipoPatNo **t, int idDoc); /* A função Insere encapsula a função InsereEntre */
 
 void imprimePatricia (TipoPatNo *no);
